Fold in dependabot, and change test structure to avoid test setup combinatorics
