+++ script/tool/lib/src/validate_command.dart
@@ -3,20 +3,39 @@
 // found in the LICENSE file.
 
 import 'package:file/file.dart';
+import 'package:meta/meta.dart';
 
 import 'common/package_looping_command.dart';
 import 'common/repository_package.dart';
+import 'validators/dependabot_validator.dart';
 import 'validators/repo_info_validator.dart';
+
+/// The set of possible validators.
+///
+/// Exposed for testing so that unit tests can target a single validator's
+/// behavior via the command without having to set everything required for
+/// every other validator to pass.
+///
+/// This is done instead of testing validators directly to ensure that testing
+/// includes things like command line parsing and run initialization.
+@visibleForTesting
+enum Validator { dependabot, repoInfo }
 
 /// A command to validate that packages follow various team conventions,
 /// guidelines, and best practices.
 ///
 /// This includes:
 /// - repository-level metadata about packages, such as repo README and
-///   auto-label entries.
+///   auto-label entries
+/// - dependabot configuration coverage
 class ValidateCommand extends PackageLoopingCommand {
   /// Creates Dependabot check command instance.
-  ValidateCommand(super.packagesDir, {super.gitDir});
+  ValidateCommand(super.packagesDir, {this.targetedValidators, super.gitDir});
+
+  /// The validators to run.
+  ///
+  /// If null, all validators are run.
+  final Set<Validator>? targetedValidators;
 
   late Directory _repoRoot;
 
@@ -26,6 +45,9 @@ class ValidateCommand extends PackageLoopingCommand {
 
   /// Packages with entries in labeler.yml.
   final Set<String> _autoLabeledPackages = <String>{};
+
+  // The set of directories covered by the repo's Dependabot configuration.
+  late DependabotCoverage _dependabotCoverage;
 
   @override
   final String name = 'validate';
@@ -44,18 +66,23 @@ class ValidateCommand extends PackageLoopingCommand {
   Future<void> initializeRun() async {
     _repoRoot = packagesDir.fileSystem.directory((await gitDir).path);
 
-    // Extract all of the repo-level README.md table entries.
-    _readmeTableEntries.addAll(
-      RepoInfoValidator.loadReadmeTableEntries(
-        repoRoot: _repoRoot,
-        packagesDir: packagesDir,
-        thirdPartyPackagesDir: thirdPartyPackagesDir,
-      ),
-    );
-    // Extract all of the labeler.yml package entries.
-    _autoLabeledPackages.addAll(
-      RepoInfoValidator.loadAutoLabeledPackages(repoRoot: _repoRoot),
-    );
+    if (_shouldRun(Validator.repoInfo)) {
+      // Extract all of the repo-level README.md table entries.
+      _readmeTableEntries.addAll(
+        RepoInfoValidator.loadReadmeTableEntries(
+          repoRoot: _repoRoot,
+          packagesDir: packagesDir,
+          thirdPartyPackagesDir: thirdPartyPackagesDir,
+        ),
+      );
+      // Extract all of the labeler.yml package entries.
+      _autoLabeledPackages.addAll(
+        RepoInfoValidator.loadAutoLabeledPackages(repoRoot: _repoRoot),
+      );
+    }
+    if (_shouldRun(Validator.dependabot)) {
+      _dependabotCoverage = DependabotValidator.loadConfig(repoRoot: _repoRoot);
+    }
   }
 
   @override
@@ -63,7 +90,12 @@ class ValidateCommand extends PackageLoopingCommand {
     final List<String> errors = [];
 
     if (package.isTopLevel) {
-      errors.addAll(await _validateRepoInfo(package));
+      if (_shouldRun(Validator.repoInfo)) {
+        errors.addAll(await _validateRepoInfo(package));
+      }
+    }
+    if (_shouldRun(Validator.dependabot)) {
+      errors.addAll(await _validateDependabot(package));
     }
 
     return errors.isEmpty
@@ -81,4 +113,24 @@ class ValidateCommand extends PackageLoopingCommand {
     );
     return validator.validatePackage(package);
   }
+
+  Future<List<String>> _validateDependabot(RepositoryPackage package) async {
+    final validator = DependabotValidator(
+      coverage: _dependabotCoverage,
+      path: path,
+      repoRoot: _repoRoot,
+      indentation: indentation,
+    );
+
+    final errors = <String>[];
+
+    errors.addAll(validator.validateDependabotCoverage(package));
+
+    // TODO(stuartmorgan): Add other ecosystem checks here as more are enabled.
+
+    return errors;
+  }
+
+  bool _shouldRun(Validator validator) =>
+      targetedValidators?.contains(validator) ?? true;
 }
