+++ script/tool/test/validate_command_dependabot_test.dart
@@ -5,7 +5,7 @@
 import 'package:args/command_runner.dart';
 import 'package:file/file.dart';
 import 'package:flutter_plugin_tools/src/common/core.dart';
-import 'package:flutter_plugin_tools/src/dependabot_check_command.dart';
+import 'package:flutter_plugin_tools/src/validate_command.dart';
 import 'package:git/git.dart';
 import 'package:test/test.dart';
 
@@ -22,10 +22,14 @@ void main() {
         configureBaseCommandMocks();
     root = packagesDir.parent;
 
-    final command = DependabotCheckCommand(packagesDir, gitDir: gitDir);
+    final command = ValidateCommand(
+      packagesDir,
+      gitDir: gitDir,
+      targetedValidators: {Validator.dependabot},
+    );
     runner = CommandRunner<void>(
-      'dependabot_test',
-      'Test for $DependabotCheckCommand',
+      'validate_dependabot_test',
+      'Test for validating dependabot coverage',
     );
     runner.addCommand(command);
   });
@@ -73,7 +77,7 @@ $gradleEntries
     createFakePackage('a_package', packagesDir);
 
     final List<String> output = await runCapturingPrint(runner, <String>[
-      'dependabot-check',
+      'validate',
     ]);
 
     expect(output, containsAllInOrder(<Matcher>[contains('No issues found!')]));
@@ -94,7 +98,7 @@ $gradleEntries
     Error? commandError;
     final List<String> output = await runCapturingPrint(
       runner,
-      <String>['dependabot-check'],
+      <String>['validate'],
       errorHandler: (Error e) {
         commandError = e;
       },
@@ -124,7 +128,7 @@ $gradleEntries
     Error? commandError;
     final List<String> output = await runCapturingPrint(
       runner,
-      <String>['dependabot-check'],
+      <String>['validate'],
       errorHandler: (Error e) {
         commandError = e;
       },
@@ -164,7 +168,7 @@ $gradleEntries
         .createSync(recursive: true);
 
     final List<String> output = await runCapturingPrint(runner, <String>[
-      'dependabot-check',
+      'validate',
     ]);
 
     expect(
@@ -197,7 +201,7 @@ $gradleEntries
           .createSync(recursive: true);
 
       final List<String> output = await runCapturingPrint(runner, <String>[
-        'dependabot-check',
+        'validate',
       ]);
 
       expect(
